+++ devtools/api-mock/contacts/index.js
@@ -5,6 +5,7 @@ import createCollectionMiddleware from '../collection-middleware';
 const actions = {
   get: createAction('Get contacts'),
   patch: createAction('Patch contact'),
+  post: createAction('Post contact'),
 };
 
 const selectors = {
@@ -20,9 +21,21 @@ const selectors = {
       throw new Error('contact not found');
     }
   ),
+  last: () => state => [...state.contacts].pop(),
+  lastLocation: () => createSelector(
+    selectors.last(),
+    contact => ({ location: `/api/v1/contacts/${contact.contact_id}` })
+  ),
 };
 
 const reducer = {
+  [actions.post]: (state, { body }) => ([
+    ...state,
+    {
+      contact_id: uuidv1(),
+      ...body,
+    },
+  ]),
   [actions.patch]: (state, { params, body }) => {
     const nextState = [...state];
     const original = state.find(contact => contact.contact_id === params.contact_id);
@@ -55,6 +68,11 @@ const routes = {
   'PATCH /:contact_id': {
     action: actions.patch,
     status: 204,
+  },
+  'POST /': {
+    action: actions.post,
+    selector: selectors.lastLocation,
+    status: 200,
   },
 };