+++ src/frontend/web_application/src/layouts/Page/components/CallToAction/components/CreateContactButton/index.js
@@ -1,9 +1,14 @@
-import { compose } from 'redux';
+import { bindActionCreators, compose } from 'redux';
+import { connect } from 'react-redux';
+import { push } from 'react-router-redux';
 import { withTranslator } from '@gandi/react-translate';
 import Presenter from './presenter';
-import { withNotification } from '../../../../../../services/notification';
+
+const mapDispatchToProps = dispatch => bindActionCreators({
+  action: () => push('/new-contact'),
+}, dispatch);
 
 export default compose(
-  withNotification(),
+  connect(null, mapDispatchToProps),
   withTranslator()
 )(Presenter);